requirements.txt
aenum==3.1.15
ai-api-client-sdk==2.2.1
ai-core-sdk==2.4.5
aiohappyeyeballs==2.4.3
aiohttp==3.11.7
aiosignal==1.3.1
altair==5.4.1
annotated-types==0.7.0
anyio==4.6.2.post1
arabic-reshaper==3.0.0
attrs==24.2.0
autots==0.6.21
awscli==1.34.15
beautifulsoup4==4.12.3
blinker==1.8.2
boto3==1.35.15
botocore==1.35.15
cachetools==5.4.0
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
chromedriver-binary-auto==0.3.1
click==8.1.7
colorama==0.4.6
contourpy==1.3.1
cycler==0.12.1
dataclasses-json==0.6.7
distro==1.9.0
dnspython==2.7.0
docutils==0.16
docx==0.2.4
docx2pdf==0.1.8
et-xmlfile==1.1.0
EZGmail==2024.8.13
fire==0.6.0
Flask==3.0.3
fonttools==4.53.1
frozenlist==1.5.0
gitdb==4.0.11
GitPython==3.1.43
google-api-core==2.19.2
google-api-python-client==2.144.0
google-auth==2.33.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.65.0
greenlet==3.1.1
gspread==5.0.0
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.27.2
idna==3.7
IMAPClient==3.0.1
imbalanced-learn==0.13.0
itsdangerous==2.2.0
Jinja2==3.1.4
jmespath==1.0.1
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kiwisolver==1.4.8
kubernetes==30.1.0
langchain==0.2.5
langchain-community==0.2.5
langchain-core==0.2.43
langchain-openai==0.1.9
langchain-text-splitters==0.2.4
langsmith==0.1.145
lxml==5.3.0
m2cgen==0.10.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.23.1
matplotlib==3.10.1
mdurl==0.1.2
metaflow==2.12.19
mlxtend==0.23.4
multidict==6.1.0
mypy-extensions==1.0.0
mysql-connector==2.2.9
narwhals==1.33.0
numpy==1.26.4
oauth2client==4.1.3
oauthlib==3.2.2
openai==1.35.3
opencv-python-headless==4.10.0.84
openpyxl==3.1.5
orjson==3.10.11
outcome==1.3.0.post0
packaging==24.2
pandas==2.2.3
patsy==1.0.1
pdf2docx==0.5.8
pillow==10.4.0
plotly==6.0.1
plotly-express==0.4.1
propcache==0.2.0
proto-plus==1.24.0
protobuf==5.28.0
pyarrow==18.0.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pydantic==2.10.1
pydantic_core==2.27.1
pydeck==0.9.1
Pygments==2.18.0
pyhumps==3.8.0
pymongo==4.10.1
PyMuPDF==1.24.9
PyMuPDFb==1.24.9
pyparsing==3.1.2
PyPDF2==3.0.1
PySocks==1.7.1
python-bidi==0.6.6
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
pytz==2024.2
pywin32==306
PyYAML==6.0.2
pyzmail36==1.0.5
rarfile==4.2
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==13.9.4
rpds-py==0.21.0
rsa==4.7.2
s3transfer==0.10.2
sap-ai-core-metaflow==1.1.21
scikit-learn==1.6.1
scipy==1.15.2
seaborn==0.13.2
selenium==4.24.0
Send2Trash==1.8.3
six==1.16.0
sklearn-compat==0.1.3
smmap==5.0.1
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
SQLAlchemy==2.0.36
statsmodels==0.14.4
streamlit==1.36.0
streamlit-chat==0.1.1
tenacity==8.5.0
termcolor==2.4.0
threadpoolctl==3.6.0
tiktoken==0.7.0
toml==0.10.2
tornado==6.4.2
tqdm==4.66.5
trio==0.26.2
trio-websocket==0.11.1
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.2
watchdog==4.0.2
websocket-client==1.8.0
Werkzeug==3.0.4
wsproto==1.2.0
yarl==1.18.0
